Step 6 a) Fix div that works like a link to an <a>
Divs that operate like links (with an onClick script) don't operate via the keyboard (can't be tabbed to, and don't handle return/space to operate).

Solution: make it a real <a> link.

Same thing is true for things that operate like buttons but are divs: just make them <button>s