Step 9: support different printing
I setup a <span> with the (#/# selected) text. Then, in CSS don't display that if we are in print.

However, noticed an issue that the useMediaQuery for narrow is true even if we are printing. Since I don't know how to specify print for the useMediaQuery, I switched to hiding the table row headers from CSS rather than from JS.